coucou

+++ corewar_dir/src/instrs/sti_calc.c
@@ -6,15 +6,15 @@
 /*   By: rghirell <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/03/06 18:47:29 by rghirell          #+#    #+#             */
-/*   Updated: 2017/03/08 09:12:26 by rghirell         ###   ########.fr       */
+/*   Updated: 2017/03/09 11:10:58 by rghirell         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "corewar.h"
 
-unsigned int	sti_calc(t_struct *env, t_process *proc, int i)
+int	sti_calc(t_struct *env, t_process *proc, int i)
 {
-	unsigned int dest;
+	short dest;
 
 	dest = char_to_int(env->map[pc_rotate(proc->pc, i)]) << 8;
 	dest = (dest + char_to_int(env->map[pc_rotate(proc->pc, i + 1)]));
